+++ src/main/java/name/joseland/mal/automation/scheduler/db/ScheduledRequestConfigRepository.java
@@ -7,6 +7,6 @@ import javax.transaction.Transactional;
 
 @Repository
 @Transactional
-public interface ScheduledRequestConfigRepository extends JpaRepository<ScheduledRequestConfig, Long> {
+public interface ScheduledRequestConfigRepository extends JpaRepository<ScheduledRequestConfig, Integer> {
 
 }
